organise lisings test and optimise imports
